Rename task repository
